Attempting to get rid of kafkaStreamsProcess jar files
