+++ kafka/kafkaStreamsEnergy/.gitignore
@@ -1,0 +1,8 @@
+.idea
+.idea/
+.DS_Store
+private.properties
+src/.DS_Store
+src/main/.DS_Store
+target/
+kafkaStreamsEnergy.iml
