--- kafka/kafkaStreamsProcess/.gitignore
@@ -1,8 +1,0 @@
-.idea
-.idea/
-.DS_Store
-private.properties
-src/.DS_Store
-src/main/.DS_Store
-target/
-kafkaStreamsProcess.iml
